+++ rocketgram/api/send_audio.py
@@ -30,6 +30,7 @@ class SendAudio(MessageResultMixin, Request):
     title: Optional[str] = None
     thumb: Optional[Union[InputFile, str]] = None
     disable_notification: Optional[bool] = None
+    protect_content: Optional[bool] = None
     reply_to_message_id: Optional[int] = None
     allow_sending_without_reply: Optional[bool] = None
     reply_markup: Optional[ALL_KEYBOARDS] = None