+++ rocketgram/api/send_contact.py
@@ -23,6 +23,7 @@ class SendContact(MessageResultMixin, Request):
     last_name: Optional[str] = None
     vcard: Optional[str] = None
     disable_notification: Optional[bool] = None
+    protect_content: Optional[bool] = None
     reply_to_message_id: Optional[int] = None
     allow_sending_without_reply: Optional[bool] = None
     reply_markup: Optional[ALL_KEYBOARDS] = None