+++ rocketgram/api/send_document.py
@@ -28,6 +28,7 @@ class SendDocument(MessageResultMixin, Request):
     caption_entities: Optional[List[MessageEntity]] = None
     disable_content_type_detection: Optional[bool] = None
     disable_notification: Optional[bool] = None
+    protect_content: Optional[bool] = None
     reply_to_message_id: Optional[int] = None
     allow_sending_without_reply: Optional[bool] = None
     reply_markup: Optional[ALL_KEYBOARDS] = None