+++ rocketgram/api/send_game.py
@@ -20,6 +20,7 @@ class SendGame(MessageResultMixin, Request):
     chat_id: int
     game_short_name: str
     disable_notification: Optional[bool] = None
+    protect_content: Optional[bool] = None
     reply_to_message_id: Optional[int] = None
     allow_sending_without_reply: Optional[bool] = None
     reply_markup: Optional[INLINE_KEYBOARDS] = None