+++ rocketgram/api/send_invoice.py
@@ -39,6 +39,7 @@ class SendInvoice(MessageResultMixin, Request):
     send_email_to_provider: Optional[bool] = None
     is_flexible: Optional[bool] = None
     disable_notification: Optional[bool] = None
+    protect_content: Optional[bool] = None
     reply_to_message_id: Optional[bool] = None
     allow_sending_without_reply: Optional[bool] = None
     reply_markup: Optional[INLINE_KEYBOARDS] = None