+++ rocketgram/api/send_location.py
@@ -25,6 +25,7 @@ class SendLocation(MessageResultMixin, Request):
     heading: Optional[int] = None
     proximity_alert_radius: Optional[int] = None
     disable_notification: Optional[bool] = None
+    protect_content: Optional[bool] = None
     reply_to_message_id: Optional[int] = None
     allow_sending_without_reply: Optional[bool] = None
     reply_markup: Optional[ALL_KEYBOARDS] = None