+++ rocketgram/api/send_media_group.py
@@ -23,6 +23,7 @@ class SendMediaGroup(Request):
     chat_id: Union[int, str]
     media: List[InputMedia]
     disable_notification: Optional[bool] = None
+    protect_content: Optional[bool] = None
     reply_to_message_id: Optional[int] = None
     allow_sending_without_reply: Optional[bool] = None