+++ rocketgram/api/send_poll.py
@@ -35,6 +35,7 @@ class SendPoll(MessageResultMixin, Request):
     close_date: Optional[datetime] = None
     is_closed: Optional[bool] = None
     disable_notification: Optional[bool] = None
+    protect_content: Optional[bool] = None
     reply_to_message_id: Optional[int] = None
     allow_sending_without_reply: Optional[bool] = None
     reply_markup: Optional[ALL_KEYBOARDS] = None