+++ rocketgram/api/send_venue.py
@@ -27,6 +27,7 @@ class SendVenue(MessageResultMixin, Request):
     google_place_id: Optional[str] = None
     google_place_type: Optional[str] = None
     disable_notification: Optional[bool] = None
+    protect_content: Optional[bool] = None
     reply_to_message_id: Optional[int] = None
     allow_sending_without_reply: Optional[bool] = None
     reply_markup: Optional[ALL_KEYBOARDS] = None