+++ rocketgram/api/send_video.py
@@ -31,6 +31,7 @@ class SendVideo(MessageResultMixin, Request):
     caption_entities: Optional[List[MessageEntity]] = None
     supports_streaming: Optional[bool] = None
     disable_notification: Optional[bool] = None
+    protect_content: Optional[bool] = None
     reply_to_message_id: Optional[int] = None
     allow_sending_without_reply: Optional[bool] = None
     reply_markup: Optional[ALL_KEYBOARDS] = None