+++ rocketgram/api/send_video_note.py
@@ -24,6 +24,7 @@ class SendVideoNote(MessageResultMixin, Request):
     length: Optional[int] = None
     thumb: Union[InputFile, str] = None
     disable_notification: Optional[bool] = None
+    protect_content: Optional[bool] = None
     reply_to_message_id: Optional[int] = None
     allow_sending_without_reply: Optional[bool] = None
     reply_markup: Optional[ALL_KEYBOARDS] = None