+++ rocketgram/api/send_voice.py
@@ -27,6 +27,7 @@ class SendVoice(MessageResultMixin, Request):
     caption_entities: Optional[List[MessageEntity]] = None
     duration: Optional[int] = None
     disable_notification: Optional[bool] = None
+    protect_content: Optional[bool] = None
     reply_to_message_id: Optional[int] = None
     allow_sending_without_reply: Optional[bool] = None
     reply_markup: Optional[ALL_KEYBOARDS] = None