+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project.
 ### Added
 - `ChatType.sender` and `ChatType.unknown` variants.
 - `RocketgramRequest403Error` exception class.
+- `ChatActionType.choose_sticker`. 
 
 ### Fixed
 - The `command` and `deeplink` filters now only catch messages with `UpdateType.message`.